Reduce API surface to test
